chore: more realistic `random` benchmark
