+++ packages/trie/test/trie/trie.spec.ts
@@ -83,7 +83,10 @@ for (const { constructor, title } of [
     )
 
     t.test('persist the root if the `persistRoot` option is `true`', async function (st) {
-      const trie = await constructor.create({ db: new LevelDB(new MemoryLevel()), persistRoot: true })
+      const trie = await constructor.create({
+        db: new LevelDB(new MemoryLevel()),
+        persistRoot: true,
+      })
 
       st.equal(await trie.db.get(ROOT_DB_KEY), null)
 
@@ -159,7 +162,10 @@ for (const { constructor, title } of [
       )
 
       // New trie with a new database so we shouldn't find a root to restore
-      const empty = await constructor.create({ db: new LevelDB(new MemoryLevel()), persistRoot: true })
+      const empty = await constructor.create({
+        db: new LevelDB(new MemoryLevel()),
+        persistRoot: true,
+      })
       st.equal(await empty.db.get(ROOT_DB_KEY), null)
 
       st.end()
